feat: 사용자 인증 기능 역할 분리
