+++ src/main/java/com/e207/woojoobook/api/user/UserService.java
@@ -4,13 +4,13 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.e207.woojoobook.api.mail.MailSender;
-import com.e207.woojoobook.api.mail.VerificationMail;
+import com.e207.woojoobook.api.user.request.EmailCodeCreateRequest;
+import com.e207.woojoobook.api.user.request.UserCreateRequest;
+import com.e207.woojoobook.api.verification.VerificationService;
+import com.e207.woojoobook.api.verification.request.VerificationMail;
 import com.e207.woojoobook.domain.user.UserMasterRepository;
 import com.e207.woojoobook.domain.user.UserSlaveRepository;
 import com.e207.woojoobook.domain.user.UserVerification;
-import com.e207.woojoobook.domain.user.UserVerificationRepository;
-import com.e207.woojoobook.domain.user.VerificationCodeUtil;
 
 import lombok.RequiredArgsConstructor;
 
@@ -18,17 +18,15 @@ import lombok.RequiredArgsConstructor;
 @Service
 public class UserService {
 
-	private final UserVerificationRepository userVerificationRepository;
 	private final UserSlaveRepository userSlaveRepository;
 	private final UserMasterRepository userMasterRepository;
 	private final PasswordEncoder passwordEncoder;
-	private final MailSender mailSender;
+	private final VerificationService verificationService;
 
 	@Transactional
 	public void join(UserCreateRequest userCreateRequest) {
 		// TODO : 예외처리
-		UserVerification userVerification = this.userVerificationRepository.findById(userCreateRequest.getEmail())
-			.orElseThrow(() -> new RuntimeException("유효하지 않는 email입니다."));
+		UserVerification userVerification = this.verificationService.findByEmail(userCreateRequest.getEmail());
 		if (!userVerification.isVerified()) {
 			throw new RuntimeException("인증되지 않은 회원입니다.");
 		}
@@ -40,14 +38,7 @@ public class UserService {
 
 	@Transactional
 	public boolean verifyEmail(VerificationMail verificationMail) {
-		// TODO : 예외처리
-		UserVerification userVerification = this.userVerificationRepository.findById(verificationMail.getEmail())
-			.orElseThrow(() -> new RuntimeException("존재하지 않는 email"));
-		userVerification.verify(verificationMail.getVerificationCode());
-
-		UserVerification save = this.userVerificationRepository.save(userVerification);
-
-		return save.isVerified();
+		return this.verificationService.verifyEmail(verificationMail);
 	}
 
 	@Transactional
@@ -57,23 +48,13 @@ public class UserService {
 			// TODO : 예외처리
 			throw new RuntimeException("사용할 수 없는 이메일입니다.");
 		}
-		String verificationCode = createVerificationUser(email);
+		String verificationCode = this.verificationService.createVerificationUser(email);
 
 		VerificationMail verificationMail = VerificationMail.builder()
 			.email(email)
 			.verificationCode(verificationCode)
 			.build();
 
-		this.mailSender.send(verificationMail);
-	}
-
-	private String createVerificationUser(String email) {
-		UserVerification userVerification = UserVerification.builder()
-			.email(email)
-			.verificationCode(VerificationCodeUtil.generate())
-			.isVerified(false)
-			.build();
-
-		return this.userVerificationRepository.save(userVerification).getVerificationCode();
+		this.verificationService.send(verificationMail);
 	}
 }
